webserver/requirements.txt
# LAN Communication Hub - Essential Dependencies

# Core web framework
Flask>=2.0.0
Flask-SocketIO>=5.0.0

# WebSocket support
python-socketio>=5.0.0
python-engineio>=4.0.0

# HTTPS/SSL support for camera/microphone access (REQUIRED)
pyOpenSSL>=23.0.0

# Additional utilities
requests>=2.25.0